refactor: Clean up hackerpack mobile menu bar
Moved the MobileDropdownMenu component into the hackerpack folder. Also redid
the component so it no longer floats or obsures any page content. Added
animations to the dropdown and reworked the way the highlighting works.

fixes #147
